refactor: clean up code and improve test structure
- Remove redundant type cast in JointFeldman.global_comms initialization
- Extract reusable MockTransport setup into a helper function
- Fix test data to properly test peers count validation

+++ src/crypto/dkg/joint_feldman.rs
@@ -165,7 +165,7 @@ impl JointFeldman {
             .expect("Partial PKs should not empty")
             .scheme();
 
-        let mut global_comms = vec![Point::zero(scheme); pks.len() as usize];
+        let mut global_comms = vec![Point::zero(scheme); pks.len()];
 
         for pks in pks.iter() {
             for (i, pk) in pks.iter().enumerate() {
